Updated classes, data types, and some minor changes.

+++ src/main/kotlin/com/kotlin/kotlincrudapp/model/Order.kt
@@ -2,7 +2,8 @@ package com.kotlin.kotlincrudapp.model
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Table
+import java.math.BigDecimal
 import java.util.UUID
 
 @Table("orders")
-data class Order(@Id var id: UUID?, val itemId: UUID, val price: Double, val amount: Double)
+data class Order(@Id var id: UUID?, val itemId: UUID, val price: BigDecimal, val amount: BigDecimal)
